bridge/requirements.txt
pyserial>=3.5
websockets>=12.0
pyinstaller>=6.0
